I have sort out the useEffect issue
